backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
sqlparse==0.5.5
